added database design
Designed a database which can be used for further progress in the project. Documented design decision for the database in addition to provided the design files as well as the initialization scripts for the database to be both used in fresh installs of the server as well as helping in the process of dockerizing the postgreSQL installation for development and potentially Production.